# go_backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smart Waste Management API
  description: |
    RESTful API for an AI & AR-Driven Smart Waste Management Platform.
    
    ## Features
    - **IoT Integration**: Real-time data from ESP32-based smart bins via MQTT
    - **User Management**: Profile and reward point tracking
    - **Driver Management**: Route optimization and task verification
    - **Dashboard**: Company management, pricing rules, and analytics
    - **Valuation Engine**: AI-based waste pricing
    
    ## Authentication
    Authentication endpoints are placeholders. Implement JWT or OAuth2 for production.
  version: 1.0.0
  contact:
    name: Smart Waste Team
    email: support@smartwaste.io
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /bins:
    get:
      tags:
        - Bins
      summary: List all bins
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of bins
    post:
      tags:
        - Bins
      summary: Register a new bin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBinRequest'
      responses:
        '201':
          description: Bin created
  /bins/needs-collection:
    get:
      tags:
        - Bins
      summary: Get bins needing collection
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
            default: 80
      responses:
        '200':
          description: Bins above threshold
...
  /bins/{id}:
    get:
      tags:
        - Bins
      summary: Get bin by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bin details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinResponse'
    put:
      tags:
        - Bins
      summary: Update bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBinRequest'
      responses:
        '200':
          description: Bin updated
    delete:
      tags:
        - Bins
      summary: Delete bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Bin deleted

  # Companies
...
components:
  schemas:
...
    CreateBinRequest:
      type: object
      required:
        - device_id
        - latitude
        - longitude
        - waste_type
        - capacity_liters
      properties:
        device_id:
          type: string
        location_name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        waste_type:
          type: string
        capacity_liters:
          type: integer
        company_id:
          type: string
          format: uuid
    UpdateBinRequest:
      type: object
      properties:
        location_name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        waste_type:
          type: string
        capacity_liters:
          type: integer
        is_active:
          type: boolean
    BinResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
        location_name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        fill_level:
          type: integer
        waste_type:
          type: string
        is_active:
          type: boolean
...